update lube (#220)
1. emake xxx.car java
implement in java, auto generate java code and JavaInit function in Cxxx.
2. emake xxx.car jni
implement in cpp, auto generate jni code.

+++ Sources/Sample/AutoGenerateJavaCodes/foregroundJava/CTestCarImpl.java
@@ -11,10 +11,10 @@ public class CTestCarImpl {
         System.loadLibrary("Your_Jni");
     }
 
-    private int mCarObj = 0;
+    private long mCarObj = 0;
 
-    public native int nativeInit();
-    public native void nativeDestroy(int carobj);
+    public native long nativeInit();
+    public native void nativeDestroy(long carobj);
 
     public CTestCarImpl() {
         mCarObj = nativeInit();
